+++ NEWS.md
@@ -1,3 +1,7 @@
 # moiraine (development version)
 
 - In `comparison_heatmap_corr()`, the user can now choose the legend position through the `legend_position` parameter.
+
+- Package here has been removed from dependencies (not needed).
+
+- `where()` function now imported from tidyselect instead of dplyr (as it required a newer version of dplyr).